# docs1/openapi.yaml
openapi: 3.1.0
info:
  title: Аким на 5 часов — Astana Quality of Life Score
  description: "Синтетическая модель хакатона: выберите ровно 5 мероприятий на бюд\
    жет до 100. Расчёт и поиск лучшего набора детерминированные; объяснение формируется LLM при настройке API или по шаблону.\
    \ Откройте POST /api/simulation/calculate, нажмите Try it out и Execute: приме\
    р уже заполнен. Score примера = 56.54307, базовый = 52.55768. Для сохранения в PostgreSQL зарегистрируйтесь через POST /api/v1/auth/register или войдите через /api/v1/auth/login, введите accessToken в BearerAuth через Authorize и используйте раздел History."
  version: v1
servers:
- url: http://localhost:8080
  description: Запуск из IDE
- url: http://localhost:8081
  description: Docker Compose с профилем postgres
...
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Зарегистрироваться и получить JWT
      description: Создаёт аккаунт с ролью USER. Email нормализуется и должен быть уникальным; адрес администратора зарезервирован. Пароль хранится как BCrypt-хеш. Ответ не кешируется.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: UserPass123!
              username: user
      responses:
        '201':
          description: Аккаунт создан, JWT выдан. Cache-Control no-store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/AuthBadRequest'
        '409':
          description: Email уже зарегистрирован или зарезервирован для администратора
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          $ref: '#/components/responses/HistoryUnavailable'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Войти по email и паролю
      description: Возвращает JWT для существующего аккаунта. Неизвестный email и неверный пароль дают одинаковый 401. Локальный администратор создаётся при первом запуске с реквизитами из ADMIN_EMAIL и ADMIN_PASSWORD.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: admin@example.com
              password: Admin
      responses:
        '200':
          description: JWT и профиль пользователя. Cache-Control no-store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/AuthBadRequest'
        '401':
          description: Неверный email или пароль
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          $ref: '#/components/responses/HistoryUnavailable'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Профиль текущего зарегистрированного пользователя
      description: Данные аккаунта из БД. Требуется действующий JWT; анонимный токен не допускается.
      operationId: me
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Профиль пользователя. Cache-Control no-store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '401':
          $ref: '#/components/responses/HistoryUnauthorized'
        '403':
          description: Анонимный профиль не является зарегистрированным аккаунтом
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          $ref: '#/components/responses/HistoryUnavailable'
...
  /api/v1/users/anonymous:
    post:
      tags: [History]
      summary: Создать анонимный профиль для истории
      description: Маршрут совместимости без тела запроса. Для новых аккаунтов используйте регистрацию и вход через JWT. Анонимный токен выдаётся один раз и действует только для истории; переноса истории в зарегистрированный аккаунт пока нет.
      operationId: createUser
      responses:
        '201':
          description: Профиль сохранён, токен выдан. Cache-Control no-store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnonymousUserResponse'
        '503':
          $ref: '#/components/responses/HistoryUnavailable'
  /api/v1/simulations:
    post:
      tags: [History]
      summary: Рассчитать и сохранить сценарий
      description: Валидирует 5 решений и сохраняет полный результат. 201 возвращается после завершения транзакции. Каждый успешный POST создаёт новую запись.
      operationId: save
      security:
      - BearerAuth: []
      - AnonymousBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationRequest'
            example:
              decisions:
              - measureId: M7
                districtId: nura
              - measureId: M8
                districtId: nura
              - measureId: M10
                districtId: nura
              - measureId: M12
              - measureId: M5
                districtId: saryarka
      responses:
        '201':
          description: Сценарий сохранён. Cache-Control no-store.
          headers:
            Location:
              description: Маршрут чтения сохранённого сценария
              schema:
                type: string
                example: /api/v1/simulations/11111111-1111-4111-8111-111111111111
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSimulationResponse'
        '400':
          $ref: '#/components/responses/HistoryBadRequest'
        '401':
          $ref: '#/components/responses/HistoryUnauthorized'
        '422':
          description: Невалидный набор решений; запись не создаётся
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblem'
        '503':
          $ref: '#/components/responses/HistoryUnavailable'
    get:
      tags: [History]
      summary: История текущего профиля
      description: Краткие результаты, сначала новые — createdAt DESC, id DESC. Владелец определяется по токену, фильтр userId не поддерживается.
      operationId: list
      security:
      - BearerAuth: []
      - AnonymousBearer: []
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
          default: 20
      - name: offset
        in: query
        schema:
          type: integer
          format: int64
          minimum: 0
          default: 0
      responses:
        '200':
          description: Страница истории, в том числе пустая. Cache-Control no-store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationHistoryPage'
        '400':
          $ref: '#/components/responses/HistoryBadRequest'
        '401':
          $ref: '#/components/responses/HistoryUnauthorized'
        '503':
          $ref: '#/components/responses/HistoryUnavailable'
  /api/v1/simulations/{id}:
    get:
      tags: [History]
      summary: Прочитать сохранённый сценарий
      description: Исходные решения и сохранённый результат без нового расчёта. Чужой и отсутствующий сценарии одинаково возвращают 404.
      operationId: find
      security:
      - BearerAuth: []
      - AnonymousBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Сохранённый сценарий. Cache-Control no-store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSimulationResponse'
        '400':
          $ref: '#/components/responses/HistoryBadRequest'
        '401':
          $ref: '#/components/responses/HistoryUnauthorized'
        '404':
          description: Сценарий не найден в текущем профиле
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          $ref: '#/components/responses/HistoryUnavailable'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Получите accessToken через POST /api/v1/auth/register или /api/v1/auth/login. JWT HS256 действует по умолчанию 3600 секунд. В Swagger вставляйте только токен без префикса Bearer.
    AnonymousBearer:
      type: http
      scheme: bearer
      description: Совместимость со старой анонимной историей. accessToken из POST /api/v1/users/anonymous, без префикса Bearer. Не подходит для auth/me и admin/me.
  responses:
    AuthBadRequest:
      description: Некорректный JSON, отсутствующие или недопустимые поля. Пароль и его хеш не возвращаются.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    HistoryBadRequest:
      description: Некорректное тело, UUID или параметры страницы
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    HistoryUnauthorized:
      description: Токен отсутствует или недействителен
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    HistoryUnavailable:
      description: Хранилище недоступно или профиль postgres выключен
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, username]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Уникальный email; пробелы по краям удаляются, регистр приводится к нижнему
          example: user@example.com
        password:
          type: string
          format: password
          writeOnly: true
          minLength: 8
          maxLength: 72
          description: Дополнительно максимум 72 байта UTF-8. Пароль не обрезается и не нормализуется.
          example: UserPass123!
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Отображаемое имя после удаления пробелов по краям
          example: user
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          example: admin@example.com
        password:
          type: string
          format: password
          writeOnly: true
          minLength: 1
          maxLength: 72
          description: Проверяется без удаления пробелов; для заданного администратора допустим пароль Admin.
          example: Admin
    AuthResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          description: JWT, подписанный HS256; передаётся как Bearer в Authorization
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          format: int64
          description: Время действия нового JWT в секундах
          example: 3600
        user:
          $ref: '#/components/schemas/AuthUserResponse'
    AuthUserResponse:
      type: object
      required: [id, email, username, role, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        createdAt:
          type: string
          format: date-time
    AnonymousUserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        accessToken:
          type: string
          description: Секретный токен профиля, выдаётся только при создании
        tokenType:
          type: string
          example: Bearer
        createdAt:
          type: string
          format: date-time
    SavedSimulationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        request:
          $ref: '#/components/schemas/SimulationRequest'
        result:
          $ref: '#/components/schemas/SimulationResult'
    SimulationHistoryPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SimulationHistoryItem'
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        hasMore:
          type: boolean
          description: Есть ещё записи после текущей страницы
    SimulationHistoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        modelVersion:
          type: string
        finalScore:
          type: number
        baselineScore:
          type: number
        scoreDelta:
          type: number
        budgetSpent:
          type: integer
          format: int32
    Decision:
      type: object
      properties:
        measureId:
          type: string
          description: Уникальный ID из GET /api/v1/measures
          example: M7
          pattern: "^M([1-9]|1[0-4])$"
        districtId:
          type: string
          description: Обязателен для районных мер. Для городских отсутствует или
            null.
          enum:
          - esil
          - almaty
          - saryarka
          - baikonur
          - nura
          example: nura
      required:
      - measureId
    SimulationRequest:
      type: object
      properties:
        decisions:
          type: array
          items:
            $ref: "#/components/schemas/Decision"
          maxItems: 5
          minItems: 5
      required:
      - decisions
    AppliedSynergy:
      type: object
      properties:
        measureIds:
          type: array
          items:
            type: string
        districtId:
          type: string
        metric:
          type: string
        bonus:
          type: number
    Budget:
      type: object
      properties:
        limit:
          type: integer
          format: int32
        spent:
          type: integer
          format: int32
        remaining:
          type: integer
          format: int32
    CriticalMetric:
      type: object
      properties:
        districtId:
          type: string
        districtName:
          type: string
        metric:
          type: string
        value:
          type: number
    DistrictResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        populationShare:
          type: number
        before:
          type: object
          additionalProperties:
            type: number
        after:
          type: object
          additionalProperties:
            type: number
        metricDeltas:
          type: object
          additionalProperties:
            type: number
        scoreBefore:
          type: number
        scoreAfter:
          type: number
        scoreDelta:
          type: number
...
    MeasureEffect:
      type: object
      properties:
        measureId:
          type: string
        name:
          type: string
        categoryId:
          type: string
        scope:
          type: string
        targetDistrictId:
          type: string
        cost:
          type: integer
          format: int32
        lagQuarters:
          type: integer
          format: int32
        realizationFactor:
          type: number
        affectedDistrictIds:
          type: array
          items:
            type: string
        realizedEffects:
          type: object
          additionalProperties:
            type: number
          description: "Эффекты в каждом затронутом районе с учётом лага, до clip\
            \ и синергий"
        districtScoreContributionBeforeClip:
          type: number
          description: Прямой вклад в D района до clip и синергий; не вклад в нелинейный
            FINAL SCORE
    ScoreSummary:
      type: object
      properties:
        dAvg:
          type: number
        dMin:
          type: number
        weakestDistrictId:
          type: string
        weakestDistrictName:
          type: string
        nCrit:
          type: integer
          format: int32
        criticalMetrics:
          type: array
          items:
            $ref: "#/components/schemas/CriticalMetric"
        cityContribution:
          type: number
        weakestDistrictContribution:
          type: number
        criticalPenalty:
          type: number
        finalScore:
          type: number
    SimulationResult:
      type: object
      properties:
        metricName:
          type: string
        modelVersion:
          type: string
        finalScore:
          type: number
          description: Точный Score без промежуточного округления
          example: 56.54307
        displayScore:
          type: number
          description: "Score, округлённый до 2 знаков для показа"
          example: 56.54
        baselineScore:
          type: number
        scoreDelta:
          type: number
        budget:
          $ref: "#/components/schemas/Budget"
        horizonQuarters:
          type: integer
          format: int32
        baseline:
          $ref: "#/components/schemas/ScoreSummary"
        summary:
          $ref: "#/components/schemas/ScoreSummary"
        districts:
          type: array
          items:
            $ref: "#/components/schemas/DistrictResult"
        measureEffects:
          type: array
          items:
            $ref: "#/components/schemas/MeasureEffect"
        synergies:
          type: array
          items:
            $ref: "#/components/schemas/AppliedSynergy"
        explanation:
          $ref: "#/components/schemas/Explanation"
        bestSolution:
          description: Глобально лучший допустимый набор для текущего каталога; null для baseline и старых сохранений
          $ref: "#/components/schemas/OptimalSolution"
        comparison:
          description: Сравнение завершённого выбора пользователя с максимумом Score
          $ref: "#/components/schemas/Comparison"
...
    ValidationProblem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
    Violation:
      type: object
      properties:
        code:
          type: string
        field:
          type: string
        message:
          type: string
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
...